# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TarotGen.io API",
        "description": "The backend for **TarotGen.io** — a PHP 8.5 / Slim 4 hybrid-REST API.\n\n- Resources are plural nouns (`/decks`, `/readings`); items are `/{id}`.\n- `GET` reads, `POST` creates (→`201`) or runs a command, `PUT` replaces,\n  `PATCH` applies a partial update / flag toggle, `DELETE` removes (→`204`).\n- State transitions that aren't a plain field-set are `POST …/{id}/{verb}`\n  (e.g. `…/finalize`, `…/approve`, `…/draw`).\n- Errors are `{ \"error\": \"…\" }`; multi-message validation is `{ \"errors\": […] }`.\n- Auth is a session cookie (`PHPSESSID`) set by the `/auth/*` endpoints —\n  there are no API keys or bearer tokens for first-party calls.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Same-origin API base path"
        }
    ],
    "paths": {
        "/account/readings": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "The current user's readings, newest first",
                "description": "All of the current user's readings, newest first.",
                "operationId": "52d06296825f7d5198b9620f6ee0c77a",
                "responses": {
                    "200": {
                        "description": "Array of readings",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Reading"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sessionCookie": []
                    }
                ]
            }
        },
        "/account/readings/{reading_id}": {
            "delete": {
                "tags": [
                    "Account"
                ],
                "summary": "Delete one of the user's readings",
                "description": "Delete one of the current user's readings.",
                "operationId": "792ad97854ffdc3eb107f4dd771de15c",
                "parameters": [
                    {
                        "name": "reading_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Reading not found"
                    }
                },
                "security": [
                    {
                        "sessionCookie": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Account"
                ],
                "summary": "Update a reading's name, notes, hidden-author flag, or view password",
                "description": "Update a reading the current user owns: name, hidden author, view password.",
                "operationId": "f2bf91c45a8fdd7232feb63c43aec389",
                "parameters": [
                    {
                        "name": "reading_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reading_name": {
                                        "type": [
                                            "string",
                                            "null"
                                        ]
                                    },
                                    "reading_notes": {
                                        "type": [
                                            "string",
                                            "null"
                                        ]
                                    },
                                    "hide_user": {
                                        "type": "boolean"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "remove_password": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The updated reading",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reading"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No changes provided"
                    },
                    "404": {
                        "description": "Reading not found"
                    },
                    "422": {
                        "description": "Password too short"
                    }
                },
                "security": [
                    {
                        "sessionCookie": []
                    }
                ]
            }
        },
        "/account": {
            "delete": {
                "tags": [
                    "Account"
                ],
                "summary": "Delete the account after re-entering the password (cascades readings)",
                "operationId": "f494bd5d0850bf24d66ae14b041e3bb3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Account deleted"
                    },
                    "403": {
                        "description": "Wrong password"
                    }
                },
                "security": [
                    {
                        "sessionCookie": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Account"
                ],
                "summary": "Update the account profile (currently the display name)",
                "description": "Partial update of the current account's profile. Currently the only\neditable field is the display name (previously the dedicated\nPUT /account/display-name).",
                "operationId": "54aecbcbefb600912a309f51b00f7e75",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "display_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The updated account",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Policy/uniqueness failure"
                    }
                },
                "security": [
                    {
                        "sessionCookie": []
                    }
                ]
            }
        },
...
        "/account/spreads/{user_spread_id}/submit": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Submit a personal spread to the public pending queue",
                "description": "Submit one of the user's personal spreads as a public spread (copies into\npending queue; the personal copy remains untouched).",
                "operationId": "b9dc0edda704ab480743c02a096bda87",
                "parameters": [
                    {
                        "name": "user_spread_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Submitted for review"
                    },
                    "404": {
                        "description": "Spread not found"
                    }
                },
                "security": [
                    {
                        "sessionCookie": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sessionCookie": {
                "type": "apiKey",
                "description": "Session cookie issued on sign-in. Sent automatically by the browser.",
                "name": "PHPSESSID",
                "in": "cookie"
            }
        }
    },
...
}